#37 [Adjustin Styles and removing leftovers]
Story: https://github.com/TFFLabs/SoundFoundry/issues/37
