+++ package.json
@@ -32,10 +32,8 @@
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
-    "angularfire2": "^4.0.0-rc.1",
     "codelyzer": "~3.0.1",
     "core-js": "^2.4.1",
-    "firebase": "^4.2.0",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
     "karma": "~1.7.0",
