+++ src/app/playbutton/playbutton.component.css
@@ -13,7 +13,7 @@
 }
 
 .grid {
-  width: 75%;
+  width: 100%;
 }
 
 .grid:after {
