+++ src/app/playlist/playlist.component.html
@@ -1,40 +1,38 @@
-<div class="playlist-container" dnd-droppable (onDropSuccess)="addSong($event)">
-	<div *ngIf="playlistService.tracks.length >= 1" dnd-droppable (onDropSuccess)="addSong($event)">
-		<md-list *ngIf="playlistService.tracks.length >= 1">
-			<md-list-item *ngFor="let track of playlistService.tracks">
-				<img md-list-avatar src="{{track.album.getSmallestImage().url}}" alt="{{track.album.name}}" />
-				<h3 md-line> {{track.name}}</h3>
-				<h4 md-line>
-					<a class="playlist-artist" mdTooltip="Artists">
-						<ng-container *ngFor="let artist of track.artists">
-							{{artist.name}}
-						</ng-container>
-					</a>
-					<a class="playlist-artist-album-divider">/</a>
-					<a class="playlist-album" mdTooltip="Album">
-						{{track.album.name}}
-					</a>
-				</h4>
-				<h3 md-line class="votes" class="votes">
-					<img *ngIf="track.voters.length > 0" src="{{track.voters[0].getSmallestImage().url}}" mdTooltip="{{track.voters[0].display_name}}" alt="{{track.voters[0].display_name}}"
-					/>
-					<div *ngIf="track.voters.length > 1">+{{track.voters.length - 1}}</div>
-				</h3>
-				<button [disabled]="!track.preview_url" md-icon-button (click)="playStopPreview(track)">
-					<md-icon *ngIf="(playlistService.previewing.id === track.id && playlistService.previewing.isPreviewing)">visibility_off</md-icon>
-					<md-icon *ngIf="(playlistService.previewing.id === track.id && !playlistService.previewing.isPreviewing) || playlistService.previewing.id !== track.id">visibility</md-icon>
-				</button>
-				<button md-icon-button (click)="toogleVote(track)">
-					<md-icon *ngIf="!track.isUpvoted(userService.user)">thumb_up</md-icon>
-					<md-icon *ngIf="track.isUpvoted(userService.user)" class="upvoted-icon">thumb_up</md-icon>
-				</button>
-				<button md-icon-button>
-					<md-icon>list</md-icon>
-				</button>
-			</md-list-item>
-		</md-list>
-	</div>
-	<div *ngIf="playlistService.tracks.length < 1" dnd-droppable (onDropSuccess)="addSong($event)" class="empty">
-		DRAG A SONG FROM SPOTIFY IN HERE!
-	</div>
+<div *ngIf="playlistService.tracks.length >= 1" class="playlist">
+	<md-list *ngIf="playlistService.tracks.length >= 1">
+		<md-list-item *ngFor="let track of playlistService.tracks">
+			<img md-list-avatar src="{{track.album.getSmallestImage().url}}" alt="{{track.album.name}}" />
+			<h3 md-line> {{track.name}}</h3>
+			<h4 md-line>
+				<a class="playlist-artist" mdTooltip="Artists">
+					<ng-container *ngFor="let artist of track.artists">
+						{{artist.name}}
+					</ng-container>
+				</a>
+				<a class="playlist-artist-album-divider">/</a>
+				<a class="playlist-album" mdTooltip="Album">
+					{{track.album.name}}
+				</a>
+			</h4>
+			<h3 md-line class="votes" class="votes">
+				<img *ngIf="track.voters.length > 0" src="{{track.voters[0].getSmallestImage().url}}" mdTooltip="{{track.voters[0].display_name}}" alt="{{track.voters[0].display_name}}"
+				/>
+				<div *ngIf="track.voters.length > 1">+{{track.voters.length - 1}}</div>
+			</h3>
+			<button [disabled]="!track.preview_url" md-icon-button (click)="playStopPreview(track)">
+				<md-icon *ngIf="(playlistService.previewing.id === track.id && playlistService.previewing.isPreviewing)">visibility_off</md-icon>
+				<md-icon *ngIf="(playlistService.previewing.id === track.id && !playlistService.previewing.isPreviewing) || playlistService.previewing.id !== track.id">visibility</md-icon>
+			</button>
+			<button md-icon-button (click)="toogleVote(track)">
+				<md-icon *ngIf="!track.isUpvoted(userService.user)">thumb_up</md-icon>
+				<md-icon *ngIf="track.isUpvoted(userService.user)" class="upvoted-icon">thumb_up</md-icon>
+			</button>
+			<button md-icon-button>
+				<md-icon>list</md-icon>
+			</button>
+		</md-list-item>
+	</md-list>
+</div>
+<div *ngIf="playlistService.tracks.length < 1" class="playlist-empty">
+	DRAG A SONG FROM SPOTIFY IN HERE!
 </div>
