+++ src/app/room/room.component.css
@@ -1,0 +1,36 @@
+.room-header {
+  width: 100%;
+  position: absolute;
+  text-align: left;
+  -webkit-box-sizing: border-box;
+  -moz-box-sizing: border-box;
+  box-sizing: border-box;
+  padding: 1% 0% 1% 1%;
+  background: lightblue;
+}
+
+.room-container {
+  width: 75%;
+  height: 90%;
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  text-align: left;
+  -webkit-box-sizing: border-box;
+  -moz-box-sizing: border-box;
+  box-sizing: border-box;
+}
+
+.room-header .description {
+  font-size: larger;
+  text-transform: capitalize;
+  font-style: italic;
+  color: seagreen;
+}
+
+.room-header .tittle {
+  font-size: xx-large;
+  text-transform: uppercase;
+  font-weight: bold;
+  color: gray;
+}
