+++ src/app/room/room.component.ts
@@ -10,4 +10,17 @@ export class RoomComponent implements OnInit {
   constructor(private playlistService: PlaylistService) {}
 
   ngOnInit() {}
+
+  addSong($event) {
+    const songUrls: string = $event.mouseEvent.dataTransfer.getData(
+      "text/plain"
+    );
+    
+    songUrls.split("\n").map(songUrl => {
+      const songId: string = songUrl.substr(songUrl.lastIndexOf("/") + 1);
+      if (songId) {
+        this.playlistService.addTrackToTrackList(songId);
+      }
+    });
+  }
 }
